Updated calls to <SOURCE>->from and <CLASS>->table to handle 'db.table' or 'table' formats
When Schema Loader 'qualify_objects' is on (and db_schema is defined)
table parameters include the database name followed by the table name
instead of the table name only. Either format is now supported.

This should be revisited. All the updated code makes sure it resolves to
'table' (removing the 'db.' prefix) to ensure maximum compatability.
However, a lot of these places might actually be better off with the
longer format. Not risking it for now, just making sure the API doesn't
change...