ClickableLinks: Logic now restricted specifically to A tags
There were cases where IE would detect an <img> tag with an 'href'
attribute (even though it really 'src') and this would cause the page to
navigate to the image as if it were a link when clicked on. Now, 'A' tag
is specifically tested, and all other tags are ignored, regardless of if
they have an 'href' attribute. This should fix a number of spooky IE nav
bugs that were actually being caused here all along.