TableSpec: required_fetch_columns in multi rel columns
Fixed bug where Multi rel columns would have no open link if the local
fk column was not fetched. Typically the fk column is the same as the
primary key which is already automatically always fetched so this would
rarely come up. Added the fk column to required_fetch_columns to
properly fix the bug and make the open link work in all circumstances