QUICK SEARCH MODE CAN NOW BE SET/SAVED ON THE CLIENT SIDE
If the new DbicLink2 attr 'allow_set_quicksearch_mode' is set to true
(default) the user is able to change the quicksearch_mode via new
sub-menu under the Quick Search button. Currently the only modes are
'like' (Normal) and 'exact' (Exact) but more could be added in the
future.

The setting is also persisted in saved searches