AutoPanel: Preliminary self-handling of load failures
In the event of a general (i.e. not server generated) failure during
content loading, AutoPanel now updates itself with an error message
instead of leaving the content area in a permanent "Loading..." state.
This will happen in the event of a client-side error, such as a timeout
or connection failure, and is visually distinct from the format of a
server-side exception (yellow warning icon instead of a red X icon in
the tab, plus other misc formatting differences...).

In the event of a timeout, a custom timeout-specific message is shown
with helpful tips to the user. For all other types of errors, a
generalized message is shown with the 'statusText' of the response
object displayed.