RapidDbic plugin: RELATIONSHIP LINKS/URL SETUP (hackish)
relationship column inline 'open' links are now automatically setup and
work, even for out-of-the-box Schema/Results without any TableSpec
configs defined. This is done in a somewhat hackish manner... module
paths are guessed (although fairly contained) and TableSpec configs for
'open_url_multi' and 'open_url' are set on each Result class.

This whole sub system will be refacted soon, and when that happens all
this code will likely not work (but also likely be irrelivent)