Possible record mix-up after UPDATE fixed (manual row ordering)
Fixed problem where records could get mixed up in the Ext Data Store
(in-memory) after update. This could then cause subsequent updates on
the same in-memory record to be applied to the wrong record.

It turns out that Ext uses only the order of rows returned after an
update operation to update its local cache, and ignores the record_pk.
Thus, if new values are not returned in the same order, records will
swap positions in the cache.

This problem is now fixed by manual iteration of ordered keys (as
supplied by the client, including proper tracking of rekeys), and
sorting the updated rows to match before returning them to the client