FIRST NUMBERED VERSION: v0.99001 - WORKING CPAN PACKAGE!
Setup Dist::Zilla and packaged RapidApp for the first time to be
installable as a standard CPAN package/dist. Updated to access the share
dir using File::ShareDir instead of via 'rapidapp' dir in the home dir.
RapidApp applications can now function without the 'rapidapp' directory
and instead access the installed distribution on the system.