TC: gave up on nested error handling and went back
The nested error handling turned out to be too buggy and complex, had
issues with caching, etc. Decided it is not important enough to keep
working on for now and went back to top level template error display.
This just means that if a template has errors because of errors in an
included template, the use cannot edit the sub-template directly and has
to open that template separately. The nested editing was just a nice to
have...