AppExplorer: new, simple 'dashboard_url' option
Instead of previous 'dashboard_class' config for loading a module to act
as the "dashboard" a much simpler 'dashboard_url' string option is now
available which will simply load the URL supplied in a tab. This setting
now makes sense with the recent improvements to AutoPanel and the new
Template::Controller

RapidDbic now passes the 'dashboard_url' config option into the auto
configured AppExplorer it sets up.

Also added new AutoPanel option 'autopanel_ignore_tabtitle' to be able
to override 'tabTitle' (and friends) supplied in the loaded panel/html