TC: moved get_template_vars() into the 'Access' class
The Template Access class now provides the TT variables used to render
the templates. This provides the API for limiting/customizing the
variables available based on both the current request/user as well as
the specific template, if desired. This is critical whenever non-admins
will be granted access to write templates.

IMPORTANT: the default Access class provides access to the context ($c)
object in [% c %] which means any user than can create templates can
access anything provided by the context API (which is nearly
everything). Obviously, this is only suitable for admins if write access
is turned on (but write access is still disabled by default)