TC: Added built-in 'admin_tpl' (& 'non_admin_tpl') Access perms
Building on the template var feature added in the previous commit,
built-in permissions have been added to distinguish between 'admin' and
'non-admin' templates. Each type get a different set of vars when they
are rendered -- admin templates get access to 'c' while non-admin
templates do not. This is important only when provide write/create
access to limited/non-admin users to protect the all-powerful context
object. Any user that has access to write to admin templates *is* an
admin because they are basically able to write whatever code they want.

As descrbed in commets in the code, here is an example of an Access
config that might be supplied to provide safe access to allow limited
users to edit certain templates:

   writable_regex      => '^wiki',
   creatable_regex     => '^wiki',
   non_admin_tpl_regex => '^wiki',

TODO: This is not yet *officially* safe and the topic of Template access
in general needs to be fully fleshed out (for example, is the RAWPERL TT
option on by default?) I am sure there are more examples. Note that this
still only matters when template editing is enabled which it is not by
default.