TC: Added 'format' meta data handling
Extended the Access API to return an arbitrary format string associated
with a given template name. Currently this is 'markdown' for *.md and
'html' for everything else. This format is then used in the Context
object for post-process parsing and is also included in the div wrap
meta tag as 'template-format' (along side the existing 'template-name')

Still in progress...