API CHANGE: Global replace 'icon-' -> 'ra-icon-'
Globally changed the CSS Icon prefix from 'icon-' to 'ra-icon-' for
RapidApp assets and CSS/icon referecnes in-code. Default app icon prefix
is still 'icon-'.

This was done to provide cleaner separation moving forward. However, all
existing apps that reference RapidApp CSS icons will have to be updated.
This could unfortunately be tricky because both app icons and RapidApp
icons were in the 'icon-' namespace (this ambiguity, in fact, is the
reason this is being done now rather than later when there are other
apps depending on it)