LOCAL LINK CLICKS NOW TRANSFORMED INTO HASHPATHS
Clicking a *local* link such as <a href="/some/path"> now navigates to
the hashpath '#!/some/path' which will open in a tab.

Relative urls are also handled. For example, if the current hashpath is
'#!/tpl/wiki/foo.tt', clicking a link with href="blah.tt" would navigate
to '#!/tpl/wiki/blah.tt'

This was done so that template links can be consistent and work
correctly in both browser and Ext/tab context. Also, AppTab was recently
improved to be able to display raw text/html content in addition to
JSON-based Ext container configs, so this is now a reasonable default
behavior