TC: ADDED 'external_tpl' ACCESS PROPERTY + handling
Templates can now be 'external' which provides different handling rules:

 * Is always loaded in an iframe tab (if inside ExtJS client)
 * Includes a limited set of css/js asset <head> includes
 * Is not wrapped in 'ra-scoped-reset' div

External templates are intended to be viewed outside the ExtJS interface
and need to be totally isolated from the ExtJS styles. This was the
original purpose of the ra-scoped-reset, but that technique is not a
100%/perfect solution like the iframe tab is.

Templates are identified as 'external' via the Access class via the new
'external_tpl' property. Like 'admin_tpl' and 'non_admin_tpl'
properties, 'external_tpl' is a property of the template, not the
current user.

TODO: external templates should also be public by default (i.e. not
require a logged in user to be accessed). This has not been done yet...