TC: ADDED template_render() TO THE TEMPLATE::CONTROLLER
This provides a new mechanism to render templates within any code rather
than during a specific Template::Controller request. Functionality has
been added to allow this to work with or without a $c object.

This is meant to replace $c->template_render() eventually. The first
place the the new TC->template_render is being used is to render the
login page in the AuthCore plugin.

Note that the current API as it stands now may change...