TC: new format 'html-snippet'
By default templates that end in *.tt have the format 'html-snippet'.
*.md is still 'markdown' and all others are 'html'. 'html-snippet' is
the same as 'html' except now only 'html-snippet' is edited within the
HtmlEditor. This is needed because the current HtmlEditor (which will
hopefully be replaced soon) has problems with content that contains
<body> so it is only really suitable for snippets

Also, <base target="_blank"> is only added in the context of a iframe
'request' (i.e. fetched by the iframe)