AuthCore: CUSTOM PUBLIC PAGES NOW SUPPORTED!
Using new config params, an arbitary set of public pages can be defined
to display instead of the login page. For example:

  'Plugin::RapidApp::AuthCore' => {
    public_template_prefix => 'public/section/',
    public_root_template => 'public/section/Home.html'
  },

The pages are based on Template::Controller templates. In the above example,
http://example.com/ would render the template 'public/section/Home.html'
and http://example.com/foo/bar.tt would render the template
'public/section/foo/bar.tt'.

When this is enabled, the only way to login is to access
http://example.com/auth/login directly, or via custom form that submits to
the login url (also at http://example.com/auth/login).

Note that this disables hashnav paths when not logged in.