ROOT_TEMPLATE_PREFIX REFACTOR: unified url paths
Refactored and unified functionality for public, external pages with URL
prefix aliases. The logic is now contained (mostly) within the Root
Module (instead of AuthCore). This allows aliases to work separately
from AuthCore config.

Root template also now works within the ExtJS/tab interface. '#!/' is
now a valid path and is now the same as '/' (from the outside)

The new API (config example) is:

  'Model::RapidApp' => {
    root_template_prefix => 'public/section/',
    root_template => 'public/section/Home.html'
  },

Currently, the root_template_* paths are always public, but this is going
to be allowed to be configured independently