AuthCore: NOW WITH CONFIGURABLE HASHED PASSWORDS
New config param 'pw_type' can be set to 'clear' (default) or a
hash_type (such as 'SHA-1', 'MD5', etc). This is a passthrough to
'password_hash_type' password/credential config.

However, this setting is also applied as expected to interfaces for
setting passwords. There is a new writable virtual column 'set_pw' in the
CoreSchema User table which sets the 'password' column with the supplied
value, passed through the configured hash function. So, if the pw_type
is 'MD5' the password will be set with MD5, if it is SHA-1 it is set
with SHA-1. Just like Plugin::Authentication, this is a direct
passthrough to Digest->new() (i.e. Digest::MD5, etc) so there is no
predetermined/limited possible hash types.

Additionally, when a fresh CoreSchema is initialized, the auto-created
admin user will have its password set through the new set_pw virtual
column. Also, a new config param 'init_admin_password' is now available
to be able to set what that initial password will be (default still
'pass').

Finally, a new bool config param 'no_validate_passwords' is available
which will set the credential password type to 'none' along with a
warning message. This allows being able to login with any password to be
able to make changes in the interface, such as changing the admin
password if the pw_type has been changed since the CoreSchema was first
initialized.