AuthCore: API CHANGE - refactored for 'Authen::Passphrase'
Default is now 'BlowfishCrypt' (cost 9), but there are new config params to
allow setting a custom type. Examples:

  'Plugin::RapidApp::AuthCore' => {
    passphrase_class => 'Authen::Passphrase::SaltedDigest',
    passphrase_params => {
      algorithm => "SHA-1",
      salt_random => 20
    }
  },

Or, for a higher bcrypt cost:

  'Plugin::RapidApp::AuthCore' => {
    passphrase_class => 'Authen::Passphrase::BlowfishCrypt',
    passphrase_params => {
      cost        => 14,
      salt_random => 1,
    }
  },

The settings only apply to 'set_pw' in the CoreSchema User table, which is
used for setting new password. Once the password is set, authenticating it
will continue to work, even if the setting is changed later. This is because
the passphrase contains its own encoding metadata.