DbicLink: Hard-coded 'sqlite_see_if_its_a_number' => 1 - fixes #10
The SQLite-specific $dbh attribute 'sqlite_see_if_its_a_number' is now
set (localized) to true in read_records() in DbicLink2. This workaround
solves the problem of multi-rel column (related counts) conditions not
working under SQLite which is discussed in the DBD::SQLite docs:

 https://metacpan.org/module/DBD::SQLite#Functions-And-Bind-Parameters

This info was forwarded by ribasushi, although he suggested a DBIC/Rs
condition level workaround which would require a lot a surgery to
implement in RapidApp, while this single flag is clean and easy to set.

This seems like a safe across the board setting for RapidApp for now.
It may need to be revisited if it causes some other unexpected issues.
Because the setting is localized it won't effect any manual/user code
queries called outside of the main DbicLink connector, so I think this
risk should be relatively low.