DbicAppCombo2: Fixed virtual display column order_by (stop-gap)
See in-code comments. This is a stop-gap fix for the regression from the
last change to DbicAppCombo2 which set the default sorting, which didn't
consider the special case of the display column being a virtual column.
This code as already been addressed in DbicLink2, and DbicAppCombo2 has
long been needing to be refactored for DbicLink2. But this is just a
practical tmp measure. I don't like it and this needs to be fixed, and
now there are lots of comments/reminders in the code saying so...