Menu fields now render with down bullet/icon - fixes #30
Expanded and improved 'menu-field' fields to always render with an icon
(in form/field context). They already supported rendering with an icon
if manually specified (which the 'bool' column profile automatically
sets up), but now, when there is no iconCls specified, a generic down
arrow bullet icon is shown instead (i.e. for the case of the new 'enum'
handling). This just looks nicer, and makes it easier to see within a
form that the value can be changed.