AutoPanel: reload link now available in comm fail tab - fixes #24
The "Communication Failure" page which is displayed upon Ajax failure
now provides links prompting the user to retry/reload. This uses a new
general support added to AutoPanel across-the-board which will trigger a
reload upon any click on an HTML element with the css class
'ra-autopanel-reloader'.

All the images used on the failure page are also now prefected (see
previous commit) to ensure the page displays correctly.