RapidDbic: 'menu_require_role' (+ TabGui 'require_role')
Added new simple role-based permissions API to allow excluding navtrees
(i.e. top-level items in the nav tree) from the TabGui according to a
simple role test, which be default uses the role table/system of
AuthCore, but can be custom defined as well.

Added top level 'menu_require_role' option to the RapidDbic config to
pass a 'require_role' to the navtree that RapidDbic sets up. Currently
this is a global on/off test, to either include all models or none of
them. The idea however is that this feature can be used with NavCore to
only allow administrators, for example, to see the source RapidDbic
grids, but still be able to grant access to non-admin users via the
Saved Search system (i.e. the only view they will be able to see are
those that have been configured as public searches).

Note that this is NOT meant to be a security feature currently, and is
instead meant to reduce clutter simply by changing the default menu
points that are presented, as the user can still access any of the pages
if they know the URL. It is still up to the individual pages/modules to
enforce the real permissions on the view and CRUD requests.