TableSpec: CODEREF RELATIONSHIPS EXCLUDED (temp)
To allow better testing of RapidApp, instead of throwing an exception
when a coderef relationship is encountered, they are simply excluded.
This is just a short-term thing for easier testing.

Hopefully this can go away sooner than later by adding support for the
coderef relationships, but this will involve a long-planned refactor...