Prelim experimental "Virtual Single Relationship Columns" support - refs #40
Single relationship columns require single-key conditions because there
must be a 1-to-1, forward and backward foreign key relationship in order
for single-selection dialogs to function. This is part of the reason
that this requirement exists...

However, single-relationships can also exist separately from proper
foreign key definitions in DBIC, and in order to support them in
RapidApp, they need to be treated differently, which this commit
provides experimental, preliminary support for, calling them "Virtual
Single Relationship Columns"

These columns are nearly the same as Multi Relationship Columns, in that
they are virtual, except the sub query returns the associted value of
display_column instead of a count.

There is no custom renderer/open link support yet, and the way this
is currently implemented is just for testing the idea...

In progress...