AppExplorer: now checks 'menu_require_role' in 'navtrees' config
The param 'menu_require_role' can now be specified as a param within the
'navtrees' config to check that the current user has the role before
including the navtree item. The works in the same manner as other places
that use 'require_role' construct...

Note that this is just for display purposes (showing the item in the
menu), hence the name 'menu_require_role' instead of just
'require_role'. Access will *not* be restricted to the actual
sub-modules if the user knows the URL path. It is assumed those modules
provide their own permissions. This same concept has already been
implemented in AppDbicTree...