Raw (uncaught) Ajax request exceptions now shown
When an exception occurs in the backend during an Ajax request, the
server-side is supposed to set the X-RapidApp-Exception header, which is
used by the client to display the exception dialog to the user. If the
header is not set for whatever reason (i.e. some server-side failure)
the client side will now display the exception dialog with the raw
response body as text. This will probably be ugly, but is better than
silently discarding the unexpected response

Note: this should still never happen in production...