Generic Ajax request failures now generate exception dialog
Generic Ajax requests - which are *not* associated with loading the
content for an AutoPanel - which fail are now shown to the user in an
error window instead of being ignored. This applies to requests which
fail, meaning the sever could not be reached or did not return a valid
response. This is different from *exceptions* which are returned from
the server in a completed request, which were already handled.

This was also already handled for AutoPanels. In this commit, it is now
handled for the other Ajax requests (such as CRUD actions)