SimpleCAS: Image::Resize (and GD) now optional - fixes #42
When oversided images are uploaded to the SimpleCAS, Image::Resize is
used to make them smaller only if it is available. If it is not
available, the image is still displayed smaller, but the original
file/size itself is not changed. The warning message displayed to the
user (i.e. in HtmlEditor) is different depending on which the case is.

This removes Image::Resize as a build/install dependency, which should
also remove GD as a dependency. This is being done because GD has a lot
of its own dependencies to get working to install on new systems and is
a hassle in general.

This may be revisited in the future to add warnings or other options to
require real Image resizing if that is what the user wants (since there
are lots of practical cases where end-users can't be trusted to insert
reasonably sized images on their own, such as huge images from their
digital camera, etc, and this can take up lots of unnecessary storage
and bandwidth, depending on the application). Another possibility to
explore instead of GD is an ImageMagick-based solution instead, since it
is much simpler to install and get working (and using GD just for image
resizing is like killing a fly with a sledgehammer).