TC: Hard-coded UTF-8 overrides
All templates within the context of the Template::Controller are now
treated as UTF-8. Template content is both encoded as UTF-8 when
templates are modified, and decoded as UTF-8 as well. BOMs of other
encodings are now ignored, because UTF-8 is the only allowed encoding
and is assumed by other places in the system.