Don't require Catalyst 5.90060 yet
There is a bug in Catalyst 5.90060 that breaks the AuthCore plugin. This
needs to be fixed in Catalyst:

 https://github.com/perl-catalyst/catalyst-runtime/pull/31

As soon as the new version is fixed/released, we will require the new
version of Catalyst (assume it will be 5.90061). In the meantime,
changing the prereq version to the last version that was required which
is 5.90042