LinkClickCatcher works again (w/o breaking browser history) - fixes #43
Browser history was "fixed" in 8733c3c3 but this was actually just
having the effect of disabling the LinkClickCatcher, which is where the
actual problem existed, which was catching clicks on tab elements
(href='#') when it shouldn't have. This was not causing a nav, but was
changing the history stack which would subsequently break proper
functioning of forward/back buttons. The link catcher now ignores hrefs
that start with #, since normal hashnav urls are already handled
normally... We only want to catch links to *local*, real paths