ADDED CONFIG OPTION 'module_root_namespace'
The "mount" point of the RapidApp root module can now be det using the
'Model::RapidApp' config key 'module_root_namespace'. By default this is
empty string ('') (i.e. '/') but can now be set to something else and
the app (such as TabGui) will operate correctly under the custom path.

This is still preliminary and not fully tested yet...