Split DirectCmp controller into 'direct' and 'navable' (unimplemented)
The default still falls back to direct, so the following are currently
equivelent:

  * /rapidapp/module/...
  * /rapidapp/module/direct/...

But there is also now

  * /rapidapp/module/navable/...

The idea being different modes of accessing/rendering a module. navable
(unimplemented/throws exception) will wrap the target module in a
navable paranet (i.e. AppTab tab panel), and additional types may be
added, also... and the default/fallback may be changed or removed - as
of now the published API is to use the full /rapidapp/module/direct/...
path...