DirectCmp 'navable' action working + improvements to RapidApp::Viewport
The 'navable' subpath in the DirectCmp controller can now be used to render
the RapidApp module path in a dedicated AppTab tabpanel which is setup
as the main-load-target. This is now an alternate to 'direct' which
still loads the module directly in the Viewport (and thus cannot handle
hashnav actions, unless of course it *is* itself a navable/apptab).

The main RapidApp::Viewport view has also been enhanced to support
either config_url/config_params, OR new 'panel_cfg' to be used to render
a static panel config outright (and this new mode is what is used by the
new 'navable' DirectCmp action, but can be used anywhere).