rapidapp.pl - RapidDbic helper creates runtime relative sqlite dsn path + misc
The RapidDbic helper trait for rapidapp.pl now calls to a special
Catalyst model helper, 'DBIC::Schema::ForRapidDbic', which adds support
to the standard 'DBIC::Schema' helper for setting up a *runtime*
relative path to the database file for the case of SQLite dsns. This
allows the connection to work consistenty for dev and deployment, not
care the shell cwd, etc.

Also, rapidapp.pl no longer creates the .conf config file like
catalyst.pl does, since we do not automatically load the ConfigLoader
plugin (of course there is nothing stopping the user from adding this
after the fact).

Also updated dist.ini to require the most recent version of Catalyst
which is now 5.90064 (changed from 5.90061)