'Model::RapidApp' configs can be supplied in 'RapidApp'
Config params that are currently consumed from 'Model::RapidApp' (such
as root_module_namespace) can now be supplied in the app config using
the simple key 'RapidApp'. This is being done as a first step to
deprecating Model::RapidApp since what that class does isn't really /
shouldn't be called a model. The new config option is for the sake of
less confusing end-user app configs (but is still actually
Model::RapidApp under the hood, but at least now the users don't have to
see it)