CSS tweak to fix scrolling issue
Changed the top-level ra-template css height from '100%' to 'auto'. This
fixes a scrolling issue when the first element in the child content has
a margin (such as the <h1> tag) that pushes the whole content down, and
makes a scrollbar appear even when there should be enough space to not
need to scroll.