Fix rand load-order bug in auto_editor_type in RapidDbic
Fixed an issue in RapidDbic which would cause the custom/grid setting
for the auto_editor_type to be sporadically ignored, falling back to
the default 'combo' setting. The problem turned out to be a load-order
issue where it would randomly be expressed according to what order the
TableSpec component was loaded and applied on each result class. Also
applied some more sane defaults for auto_editor_type to take 'grid' to
mean the custom, windowed version, which is all the user would expect,
rather than a one-off grid generated on-the-fly, which should only
happen outside of RapidDbic when a pre-existing grid is not available