QUICK SEARCH - IMPROVED SUPPORT FOR NUM COLS (Pg safe) - fixes #52
Numeric columns are now handled differently than text columns for quick
search. They are forced to 'exact' mode (since it would never make
sense to attempt a LIKE/sub-string on a number) and are also proactively
skipped unless the search query is a valid number (with support for both
integer and decimal types). This is both more robust in general (and also
prevents an exception thrown by postgres, which is the main reason for
the feature).

There is also a new TableSpec column property 'quick_search_exact_only'
which can be used to force/override any column to always use 'exact' mode
(see POD).