VIRTUAL COLUMN FILTER CONDITIONS REWORKED - fixes #51
Virtual columns conditions (setup via MultiFilters) are now built
using their sub-select within a WHERE clause, instead of via alias
identifier in a HAVING clause. The reasons for this are explained and
discussed in detail in both the comments in Issue #51 as well as the
code. This was made possible by the hack function _binary_op_fuser,
provided by ribasushi.

Note that the code is not in a final state, as that will involve
factoring out a bunch of un-needed code. Not doing this yet to make
sure this really is what we want/need...