depth-limited 'dump_these' ovr excludes Request/Response - fixes #54
The RapidApp-specific 'dump_these' override which is designed to limit
the size was breaking file uploads and other places which expect to get
the original Request/Response objects out of the dump. To handle this, a
special exclusion has been added that ensures the originals for Request
and Response are preserved (but still performs a depth-limit munge on
the rest of it).

See in-code comments for addtional info/discussion...