UNFACTORED SimpleCAS INTO AN EXTERNAL PACKAGE - refs #62
The SimpleCAS controller is now loaded from the external package/class
Catalyst::Controller::SimpleCAS which is in its own repo. This will be
on CPAN eventually, but, as of this RapidApp commit, this corresponding
commit for Catalyst::Controller::SimpleCAS works the same as it did:

https://github.com/vanstyn/Catalyst-Controller-SimpleCAS/commit/e708eb52

There shouldn't be any further RapidApp-side work besides requiring the
real version number of the new package as a dependency which will happen
after it is officially released (and the issue will resolved at that
time). There is still work/cleanup to do in the new repo which will be
tracked there.