DbicAppCombo2: REFACTOR + new user_editable/type_ahead opts
DbicAppCombo2 has finally been refactored and cleaned up to behave in a
sane and straightforward manner, including proper handling of virtual
display columns (fetch in a single query).

Additionally, new 'user_editable' and 'type_ahead' options have been
added which changes the behavior to allow type to search. These options
are still experimental and are likely to change. The plan is to tie in
additional auto-configuration via auto_editor_type + TDB column profiles
to set these options, but this work is not yet done...

There have also been some more fixes and handling of edge cases in the
AppCombo JS code to work correctly in various different modes and
contexts, including grids, AppDV, add/edit form and MultiFilter.

This combo code is now very solid compared to what it was, and should be
a lot easier to maintain ging forward