MASSIVE IMPROVEMENTS: auto_editor_type => 'combo'
The 'combo' auto_editor_type now enables the brand new 'type_filter'
option in DbicAppCombo2 which enables typing in the field to
search/filter the dropdown list. The previous behavior of a
simple dropdown list can still be accessed via the newly added
auto_editor_type 'dropdown'.

There is also a new column profile 'soft_rel' which enables the
'user_editable' option which is also newly available in DbicAppCombo2
for both 'combo' and 'dropdown' auto_editor_types. The idea behind this
is to flag a single relationship column as a non-foreign key, meaning
values do not have to exist in the remote table to be valid. Later on
this profile may be automatically enabled when the DBIC rel attr
'is_foreign_key_constraint' is defined and set to false, which has
exactly the same meaning and is already used by DBIC to decide whether
or not to deploy a given single relationship as a real constraint when
generating DDL. But, for now, the soft_rel profile is available but has
to be set manually.

These new features have also been documented in the manual POD