Bad module dispatch paths render 404 page - fixes #73
Module dispatch finally handles not-found paths in a more sane manner,
returning a 404/not found page instead of throwing an exception.

This is being done via new View::RapidApp::Template which is a thin
wrapper to dispatch to the Template::Controller, but exposed via the
standard View::TT API which is to set {template} in the stash.

Note that there are other, old/legacy code, such as HttpStatus view
which is old and should be removed, but is not being done yet because
I'm not 100% sure yet