AppDbicTree now gens module names w/ source not table - fixes #72,#71
Now the full RapidDbic-generation process correctly uses the source name
instead of the table name for module paths, links, etc. This is how it
should have been all along, however, the original thinking behind the
table name was that it would be the more descriptive identifier,
however, this is just not practical from this level of abstraction since
DBIC is the real authority, whether the end-user knows it or not. One
possible problem with this current design, however, is case-collisions
(i.e. if both sources named SomeSource and Somesource exist). However,
this case should just throw an exception on startup, and is probably a
mistake anyway.